fix overflow in ed097oc4 timing LUT

+++ components/epd_driver/epd_driver.c
@@ -25,18 +25,18 @@ uint32_t skipping;
 #if defined(CONFIG_EPD_DISPLAY_TYPE_ED097OC4) ||                               \
     defined(CONFIG_EPD_DISPLAY_TYPE_ED060SC4)
 /* 4bpp Contrast cycles in order of contrast (Darkest first).  */
-const uint8_t contrast_cycles_4[15] = {30, 30, 20, 20, 30,  30,  30, 40,
-                                       40, 50, 50, 50, 100, 200, 300};
+const int contrast_cycles_4[15] = {30, 30, 20, 20, 30,  30,  30, 40,
+                                   40, 50, 50, 50, 100, 200, 300};
 
-const uint8_t contrast_cycles_4_white[15] = {30, 30, 20, 20, 30,  30,  30, 40,
-                                             40, 50, 50, 50, 100, 200, 300};
+const int contrast_cycles_4_white[15] = {30, 30, 20, 20, 30,  30,  30, 40,
+                                         40, 50, 50, 50, 100, 200, 300};
 
 #elif defined(CONFIG_EPD_DISPLAY_TYPE_ED097TC2)
-const uint8_t contrast_cycles_4[15] = {15, 8,  8,  8,  8,  8,   10, 10,
-                                       10, 10, 20, 20, 50, 100, 200};
+const int contrast_cycles_4[15] = {15, 8,  8,  8,  8,  8,   10, 10,
+                                   10, 10, 20, 20, 50, 100, 200};
 
-const uint8_t contrast_cycles_4_white[15] = {7, 8, 8, 6, 6, 6,  6,  6,
-                                             6, 6, 6, 8, 8, 10, 100};
+const int contrast_cycles_4_white[15] = {7, 8, 8, 6, 6, 6,  6,  6,
+                                         6, 6, 6, 8, 8, 10, 100};
 #else
 #error "no display type defined!"
 #endif
@@ -469,7 +469,7 @@ void IRAM_ATTR provide_out(OutputParams *params) {
 
 void IRAM_ATTR feed_display(OutputParams *params) {
   Rect_t area = params->area;
-  const uint8_t *contrast_lut = contrast_cycles_4;
+  const int *contrast_lut = contrast_cycles_4;
   switch (params->mode) {
   case WHITE_ON_WHITE:
   case BLACK_ON_WHITE:
